[Front MCP] Test enrichment without casts
